Rename common VST interface, normalise class names across versions
