+++ JammaLib/src/engine/Loop.cpp
@@ -856,7 +856,7 @@ void Loop::UnloadVstPlugin(size_t index)
 	_changesMade = true;
 }
 
-std::shared_ptr<vst::IAnyVstPlugin> Loop::GetVstPlugin(size_t index) const
+std::shared_ptr<vst::IVstPlugin> Loop::GetVstPlugin(size_t index) const
 {
 	auto chain = _vstChain.load(std::memory_order_acquire);
 	if (!chain)
