+++ JammaLib/src/engine/LoopTake.h
@@ -138,7 +138,7 @@ namespace engine
 		void SetSampleRate(float sampleRate);
 		float GetSampleRate() const noexcept { return _sampleRate; }
 		unsigned int GetLastBufSize() const noexcept { return _lastBufSize; }
-		std::shared_ptr<vst::IAnyVstPlugin> GetVstPlugin(size_t index) const;
+		std::shared_ptr<vst::IVstPlugin> GetVstPlugin(size_t index) const;
 		std::vector<io::JamFile::VstEntry> VstEntries() const;
 
 		void Record(std::vector<unsigned int> channels, std::string stationName, std::vector<unsigned int> midiChannels = {});
