+++ JammaLib/src/engine/Scene.cpp
@@ -6,7 +6,7 @@
 #include "../io/TextReadWriter.h"
 #include "../utils/PathUtils.h"
 #include "../graphics/VstEditorWindow.h"
-#include "../vst/VstPlugin.h"
+#include "../vst/Vst3Plugin.h"
 
 using namespace base;
 using namespace actions;
@@ -218,7 +218,7 @@ void Scene::_PruneClosedVstEditorWindows()
 	}), _vstEditorWindows.end());
 }
 
-bool Scene::_OpenVstEditorForPlugin(const std::shared_ptr<vst::IAnyVstPlugin>& plugin)
+bool Scene::_OpenVstEditorForPlugin(const std::shared_ptr<vst::IVstPlugin>& plugin)
 {
 	if (!plugin || !plugin->IsLoaded())
 		return false;
@@ -947,7 +947,7 @@ void Scene::OnJobTick(Time curTime)
 	// Hand any PreInit'd VST plugin (created on the UI thread) back to the UI
 	// thread for destruction. On success the chain holds its own ref so this
 	// queued ref is a no-op; on failure (Load returned false) this is the only
-	// remaining ref, and draining it on the UI thread ensures ~VstPlugin →
+	// remaining ref, and draining it on the UI thread ensures ~Vst3Plugin →
 	// IComponent::terminate() / FreeLibrary run on the thread that PreInit'd
 	// the plugin. Releasing on this job thread instead violates VST3 threading
 	// and can crash plugins or leave dangling state until window close.
@@ -1179,7 +1179,7 @@ void Scene::CommitChanges()
 	// Pre-initialise VST DLLs on the UI thread before handing jobs to the job
 	// thread. Do this after releasing _audioMutex so LoadLibraryW stays out of
 	// the audio lock and later attached() calls remain UI-thread bound.
-	// MakePluginForPath selects VST3 (VstPlugin) or VST2 (Vst2Plugin) by
+	// MakePluginForPath selects VST3 (Vst3Plugin) or VST2 (Vst2Plugin) by
 	// file extension (.dll → VST2, anything else → VST3).
 	for (auto& job : jobList)
 	{
