+++ JammaLib/src/vst/Vst.cpp
@@ -1,2 +1,2 @@
-// Legacy stub — no implementation needed; all VST3 logic is in VstPlugin.cpp and VstChain.cpp.
+// Legacy stub — no implementation needed; all VST3 logic is in Vst3Plugin.cpp and VstChain.cpp.
